requirements.txt
Flask==0.12.2
Flask-PyMongo==0.5.1
gunicorn==19.4.5
pymongo==3.6.1
PyYAML==3.12
nose==1.3.7